+++ modules/Attendance/includes/AttendanceCodes.fnc.php
@@ -26,7 +26,7 @@ function MakeAttendanceCode( $state_code, $name = '', $title = '' )
 		// Attendance codes.
 		'P' => _( 'Present' ),
 		'A' => _( 'Absent' ),
-		'H' => _( 'Half Day' ),
+		'H' => _( 'Half' ),
 		// Daily attendance.
 		'1.0' => _( 'Present' ),
 		'0.0' => _( 'Absent' ),